backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
tensorflow==2.15.0
opencv-python==4.9.0.80
librosa==0.10.1
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
supabase==2.3.4
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic-settings==2.1.0
